Moved bot.py and botconfig.py contents to __init__.py.
